add utils module and suppress rasterio's NotGeoreferencedWarning
