+++ simple_ortho/utils.py
@@ -1,0 +1,44 @@
+"""
+   Copyright 2023 Dugal Harris - dugalh@gmail.com
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+"""
+import warnings
+from contextlib import contextmanager
+from typing import Tuple, Union
+import numpy as np
+from rasterio.errors import NotGeoreferencedWarning
+from rasterio.windows import Window
+
+
+@contextmanager
+def suppress_no_georef() -> None:
+    """ Context manager to suppress rasterio's NotGeoreferencedWarning. """
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore', category=NotGeoreferencedWarning)
+        yield
+
+
+def expand_window_to_grid(win: Window, expand_pixels: Tuple[int, int] = (0, 0)) -> Window:
+    """ Expand rasterio window extents to the nearest whole numbers. """
+    col_off, col_frac = np.divmod(win.col_off - expand_pixels[1], 1)
+    row_off, row_frac = np.divmod(win.row_off - expand_pixels[0], 1)
+    width = np.ceil(win.width + 2 * expand_pixels[1] + col_frac)
+    height = np.ceil(win.height + 2 * expand_pixels[0] + row_frac)
+    exp_win = Window(col_off.astype('int'), row_off.astype('int'), width.astype('int'), height.astype('int'))
+    return exp_win
+
+
+def nan_equals(a: Union[np.ndarray, float], b: Union[np.ndarray, float]) -> np.ndarray:
+    """ Compare two numpy objects a & b, returning true where elements of both a & b are nan. """
+    return (a == b) | (np.isnan(a) & np.isnan(b))
